Move population of test directories into utils
